Perf toolkit -- integrated kernel test results comparison fixed
In case the first count in results is saturated (all results have the same count), we are not guaranteed by GENIE to get the same set of result idx, thereby comparison will likely result in marking the results as wrong. This has been fixed to not compare idxs in case the first results count is saturated.